Flatten places file structure
