+++ server/src/modules/places/http/place-router.ts
@@ -1,5 +1,5 @@
 import { Router } from 'express'
-import { GetAllPlacesController } from '../../../application/use-cases/get-all-places/get-all-places-controller'
+import { GetAllPlacesController } from '../use-cases/get-all-places/get-all-places-controller'
 
 class PlaceRouter {
   static using(getAllPlacesController: GetAllPlacesController): Router {
