--- server/src/modules/places/infra/repos/place-repo.ts
@@ -1,7 +1,0 @@
-import { Place } from '../../domain/entities/place'
-import { Result } from '../../../../shared/core/result'
-import { AppError } from '../../../../shared/core/app-error'
-
-export abstract class PlaceRepo {
-  abstract getAllPlaces(): Promise<Result<Array<Place>, AppError.UnexpectedError>>
-}
