+++ server/src/setup/application/controllers.ts
@@ -1,8 +1,8 @@
-import { UseCases, Controllers } from './types';
-import { CreateUserController } from '../../modules/users/application/use-cases/create-user/create-user-controller';
-import { GetAllClubsController } from '../../modules/legacy/clubs/application/use-cases/get-all-clubs/get-all-clubs-controller';
-import { GetAllEventsController } from '../../modules/legacy/events/application/use-cases/get-all-events/get-all-events-controller';
-import { GetAllPlacesController } from '../../modules/places/application/use-cases/get-all-places/get-all-places-controller';
+import { UseCases, Controllers } from './types'
+import { CreateUserController } from '../../modules/users/application/use-cases/create-user/create-user-controller'
+import { GetAllClubsController } from '../../modules/legacy/clubs/application/use-cases/get-all-clubs/get-all-clubs-controller'
+import { GetAllEventsController } from '../../modules/legacy/events/application/use-cases/get-all-events/get-all-events-controller'
+import { GetAllPlacesController } from '../../modules/places/use-cases/get-all-places/get-all-places-controller'
 
 export const setupControllers = (useCases: UseCases): Controllers => {
   return {
@@ -10,5 +10,5 @@ export const setupControllers = (useCases: UseCases): Controllers => {
     getAllClubs: new GetAllClubsController(useCases.getAllClubs),
     getAllEvents: new GetAllEventsController(useCases.getAllEvents),
     getAllPlaces: new GetAllPlacesController(useCases.getAllPlaces),
-  };
-};
+  }
+}
