+++ server/src/setup/application/types.ts
@@ -1,5 +1,5 @@
-import { GetAllPlacesUseCase } from './../../modules/places/application/use-cases/get-all-places/get-all-places-use-case';
-import { GetAllPlacesController } from './../../modules/places/application/use-cases/get-all-places/get-all-places-controller';
+import { GetAllPlacesUseCase } from '../../modules/places/use-cases/get-all-places/get-all-places-use-case'
+import { GetAllPlacesController } from '../../modules/places/use-cases/get-all-places/get-all-places-controller'
 import { CreateUserUseCase } from '../../modules/users/application/use-cases/create-user/create-user-use-case'
 import { GetAllClubsUseCase } from '../../modules/legacy/clubs/application/use-cases/get-all-clubs/get-all-clubs-use-case'
 import { GetAllEventsUseCase } from '../../modules/legacy/events/application/use-cases/get-all-events/get-all-events-use-case'
