+++ server/src/setup/database/mock-db.ts
@@ -1,4 +1,4 @@
-import { MockPlaceRepo } from './../../modules/places/infra/repos/implementations/mock-place-repo';
+import { MockPlaceRepo } from '../../modules/places/repos/mock-place-repo'
 import { Repos, DB } from './types'
 import { MockClubRepo } from '../../modules/legacy/clubs/infra/repos/implementations/mock-club-repo'
 import { MockEventRepo } from '../../modules/legacy/events/infra/repos/implementations/mock-event-repo'
@@ -6,7 +6,7 @@ import { MockUserRepo } from '../../modules/users/infra/repos/implementations/mo
 import { ClubEntity } from '../../shared/infra/db/entities/legacy/club.entity'
 import { EventEntity } from '../../shared/infra/db/entities/legacy/event.entity'
 import { UserEntity } from '../../shared/infra/db/entities/legacy/user.entity'
-import { PlaceEntity } from '../../shared/infra/db/entities/places/place.entity';
+import { PlaceEntity } from '../../shared/infra/db/entities/places/place.entity'
 
 interface MockEntities {
   clubs?: Array<ClubEntity>
@@ -27,7 +27,7 @@ const setupMockRepos = (entities: MockEntities): MockRepos => {
     event: new MockEventRepo(entities.events),
     user: new MockUserRepo(entities.users),
     place: new MockPlaceRepo(entities.places),
-  };
+  }
 }
 
 interface MockDB extends DB {
